Table layout in CSS (issue #33) and others
Fixed the problem in presenting huge tables in Springer CSS - e.g.,
such tables go out of the content border of the page - by using the
property “word-break” in CSS of the table. I’m aware this is a bit
strong as behaviour, but it is the only.

I’ve also reduced the font size of the header cells in table when
printing a Web-based document.

Finally, the paper about RASH to submit to sepublica has been corrected
in a few of passages.