New ODT XSLT and testbeds
A new version of the ODT to RASH XSLT file has been developed in order
to be more flexible and easy to use.

In addition, a full set of 12 ODT documents (and related sources, i.e.,
the unzipped directories of the original ODT files) has been added in
order to test the XSLT developed.

A small bug fix has been introduced in the rash.js.